backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
langchain-core>=0.3,<0.4
langchain-openai>=0.2,<0.3
langgraph>=0.2,<0.4
httpx>=0.27,<1.0
neo4j>=5.14.1

python-multipart
pypdf
python-docx
rapidfuzz
hypernetx
sqlalchemy 
fastjsonschema